修复报错函数内上边界裁切错误的bug
优化了报错输出，使用vprintf代理变参输出，可以增加更多额外信息

+++ NBT_Node.hpp
@@ -11,7 +11,7 @@
 class NBT_Node
 {
 public:
-	enum NBT_TAG : size_t
+	enum NBT_TAG : uint8_t
 	{
 		TAG_End = 0,	//½áÊøÏî
 		TAG_Byte,		//int8_t
